rmbg-ai/requirements.txt
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
fastapi
uvicorn